Refactored expense form
